requirements.txt
anyio >= 3.4.0, < 4.0.0
asyncssh_unofficial>=0.9.2
httpx>=0.25.0
pathspec>=0.11.2
prefect>=2.10.0
py2exe>=0.13.0.0
pydantic>=1.10.5
regex>=2022.10.31
setuptools>=67.4.0
sshfs>=2023.1.0
tomli>=2.0.1
typing_extensions>=4.7.1